% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxph.R
\name{coxModule}
\alias{coxModule}
\title{coxModule: Module for cox model}
\usage{
coxModule(input, output, session, data, data_label,
  data_varStruct = NULL, nfactor.limit = 10, design.survey = NULL,
  default.unires = T, id.cluster = NULL)
}
\arguments{
\item{input}{input}

\item{output}{output}

\item{session}{session}

\item{data}{reactive data}

\item{data_label}{reactuve data_label}

\item{data_varStruct}{reactive data_varStruct, Default: NULL}

\item{nfactor.limit}{nlevels limit in factor variable, Default: 10}

\item{design.survey}{reactive survey data. default: NULL}

\item{default.unires}{Set default independent variables using univariate analysis.}

\item{id.cluster}{reactive cluster variable if marginal cox model, Default: NULL}
}
\value{
regressModule
}
\description{
Module for cox model
}
\details{
DETAILS
}
\examples{
 #EXAMPLE1
}
