% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxph.R
\name{coxUI}
\alias{coxUI}
\title{coxUI: shiny modulde UI for Cox's model.}
\usage{
coxUI(id)
}
\arguments{
\item{id}{id}
}
\value{
coxUI
}
\description{
Shiny modulde UI for Cox's model.
}
\details{
Shiny modulde UI for Cox's model.
}
\examples{
 coxUI(1)
}
