% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csvFileInput.R
\name{csvFileInput}
\alias{csvFileInput}
\title{csvFileInput: Shiny module UI for file upload.}
\usage{
csvFileInput(id, label = "csv/xlsx file")
}
\arguments{
\item{id}{id}

\item{label}{label, Default: 'csv/xlsx file'}
}
\value{
Shiny UI
}
\description{
Shiny module UI for file(csv or xlsx) upload.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 library(shiny)
 ui <- fluidPage(
   sidebarLayout(
     sidebarPanel(
       csvFileInput("datafile", "Upload data (csv/xlsx format)")
     ),
     mainPanel(
       dataTableOutput("table")
     )
   )
 )
 server <- function(input, output, session) {
   datafile <- callModule(csvFile, "datafile")

   output$table <- renderDataTable({
     datafile()
   })
 }

 shinyApp(ui, server)
 }
}
}
