% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilePsInput.R
\name{FilePs}
\alias{FilePs}
\title{FilePs: Shiny module Server for propensity score analysis.}
\usage{
FilePs(input, output, session, nfactor.limit = 20)
}
\arguments{
\item{input, output, session}{Standard Shiny server parameters.}

\item{nfactor.limit}{An integer, the threshold for unique values.}
}
\value{
A reactive expression returning a list with matched data and other
information.
}
\description{
Server-side logic for propensity score analysis. It uses
`DataManager` for common data tasks and adds specific controls and
calculations for propensity score matching.
}
