% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileRepeatedInput.R
\name{FileRepeated}
\alias{FileRepeated}
\title{FileRepeated: Server for repeated measures analysis.}
\usage{
FileRepeated(input, output, session, nfactor.limit = 20)
}
\arguments{
\item{input, output, session}{Standard Shiny server parameters.}

\item{nfactor.limit}{An integer, the threshold for unique values.}
}
\value{
A reactive list with the processed `data`, `label`, and `id.gee`.
}
\description{
Server module for repeated measures analysis. It uses
`DataManager` and adds a control for selecting the repeated measures variable.
}
