% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileSurveyInput.R
\name{FileSurvey}
\alias{FileSurvey}
\title{FileSurvey: Server for survey data analysis.}
\usage{
FileSurvey(input, output, session, nfactor.limit = 20)
}
\arguments{
\item{input, output, session}{Standard Shiny server parameters.}

\item{nfactor.limit}{An integer, the threshold for unique values.}
}
\value{
A reactive list with `data`, `label`, `naomit`, and the `survey` object.
}
\description{
Server module for survey data analysis. It uses `DataManager`
and adds controls and logic for creating a `survey.design` object.
}
