% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestcox.R
\name{forestcoxServer}
\alias{forestcoxServer}
\title{forestcoxServer:shiny module server for forestcox}
\usage{
forestcoxServer(
  id,
  data,
  data_label,
  data_varStruct = NULL,
  nfactor.limit = 10,
  design.survey = NULL
)
}
\arguments{
\item{id}{id}

\item{data}{Reactive data}

\item{data_label}{Reactive data label}

\item{data_varStruct}{Reactive List of variable structure, Default: NULL}

\item{nfactor.limit}{nlevels limit in factor variable, Default: 10}

\item{design.survey}{reactive survey data. default: NULL}
}
\value{
Shiny module server for forestcox
}
\description{
Shiny module server for forestcox
}
\details{
Shiny module server for forestcox
}
\examples{

library(survival);library(dplyr);library(jstable);library(shiny)
lung \%>\%
  mutate(
    status = as.integer(status == 1),
    sex = factor(sex),
    kk = factor(as.integer(pat.karno >= 70)),
    kk1 = factor(as.integer(pat.karno >= 60))

  ) -> lung
out<-lung
ui <- fluidPage(
  sidebarLayout(
    sidebarPanel(
      forestcoxUI('Forest')
    ),
    mainPanel(
      DTOutput('tablesub')
    )
  )
)


server <- function(input, output, session) {
  data<-reactive(out)
 label<-reactive(mk.lev(out))
 outtable<-forestcoxServer('Forest',data=data,data_label=label)
  output$tablesub<-renderDT({
    outtable()
  })
}

shinyApp(ui, server)

}
\seealso{
\code{\link[forestploter]{forest}}
 \code{\link[rvg]{dml}}
 \code{\link[officer]{read_pptx}}, \code{\link[officer]{add_slide}}, \code{\link[officer]{ph_with}}, \code{\link[officer]{ph_location}}
}
