% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestcox.R
\name{forestcoxUI}
\alias{forestcoxUI}
\title{forestcoxUI:shiny module UI for forestcox}
\usage{
forestcoxUI(id, label = "forestplot")
}
\arguments{
\item{id}{id}

\item{label}{label, Default: 'forestplot'}
}
\value{
Shinymodule UI
}
\description{
Shiny module UI for forestcox
}
\details{
Shinymodule UI for forestcox
}
\examples{

library(survival);library(jstable);library(shiny);library(DT)

lung$status<-factor(as.integer(lung$status == 1))
lung$sex<-factor(lung$sex)
lung$kk<-factor(as.integer(lung$pat.karno >= 70))
lung$kk1<-factor(as.integer(lung$pat.karno >= 60))

out<-lung
ui <- fluidPage(
  sidebarLayout(
    sidebarPanel(
      forestcoxUI('Forest')
    ),
    mainPanel(
      DTOutput('tablesub')
    )
  )
)


server <- function(input, output, session) {
  data<-reactive(out)
 label<-reactive(mk.lev(out))
 outtable<-forestcoxServer('Forest',data=data,data_label=label)
  output$tablesub<-renderDT({
    outtable()
  })
}

shinyApp(ui, server)

}
