% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestglm.R
\name{forestglmUI}
\alias{forestglmUI}
\title{forestglmUI:Shiny module UI for forestglm}
\usage{
forestglmUI(id, label = "forestplot")
}
\arguments{
\item{id}{id}

\item{label}{label, Default: 'forestplot'}
}
\value{
Shinymodule UI
}
\description{
Shiny module UI for forestcox
}
\details{
Shinymodule UI for forestglm
}
\examples{
\dontrun{
if(interactive()){
library(survival);library(jstable);library(data.table);library(shiny);library(DT)

lung$status<-factor(as.integer(lung$status == 1))
lung$sex<-factor(lung$sex)
lung$kk<-factor(as.integer(lung$pat.karno >= 70))
lung$kk1<-factor(as.integer(lung$pat.karno >= 60))


ui <- fluidPage(
  sidebarLayout(
    sidebarPanel(
      forestglmUI('Forest')
    ),
    mainPanel(
      DTOutput('tablesub')
    )
  )
)

out<-lung

server <- function(input, output, session) {
  data<-reactive(out)
  label<-reactive(mk.lev(out))
  outtable<-forestglmServer('Forest',data=data,data_label=label,family='gaussian')
  output$tablesub<-renderDT({
    outtable()
  })
}

shinyApp(ui, server)

 }
}
}
