% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gee.R
\name{GEEModuleLinear}
\alias{GEEModuleLinear}
\title{GEEModuleLinear: Module for linear GEE using reactive data}
\usage{
GEEModuleLinear(input, output, session, data, data_label,
  data_varStruct = NULL, nfactor.limit = 10, id.gee)
}
\arguments{
\item{input}{input}

\item{output}{output}

\item{session}{session}

\item{data}{reactive data, ordered by id.}

\item{data_label}{reactive data_label}

\item{data_varStruct}{data_varStruct, Default: NULL}

\item{nfactor.limit}{nlevels limit in factor variable, Default: 10}

\item{id.gee}{reactive repeated measure variable}
}
\value{
GEEModuleLinear
}
\description{
Module for linear GEE using reactive data
}
\details{
Module for linear GEE using reactive data
}
\examples{
 #EXAMPLE1
}
