% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gee.R
\name{GEEModuleUI}
\alias{GEEModuleUI}
\title{GEEModuleUI: ModuleUI for GEE}
\usage{
GEEModuleUI(id)
}
\arguments{
\item{id}{id}
}
\value{
GEEModuleUI
}
\description{
ModuleUI for GEE
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
