% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpairs.R
\name{ggpairsModuleUI2}
\alias{ggpairsModuleUI2}
\title{ggpairsModuleUI2: Option & download module UI for ggpairs}
\usage{
ggpairsModuleUI2(id)
}
\arguments{
\item{id}{id}
}
\value{
ggpairsModuleUI2
}
\description{
Option & download module UI for ggpairs
}
\details{
DETAILS
}
\examples{
 #EXAMPLE1
}
