% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsBasicGadget.R
\name{jsBasicAddin}
\alias{jsBasicAddin}
\title{jsBasicAddin: Rstudio addin of jsBasicGadget}
\usage{
jsBasicAddin()
}
\value{
Rstudio addin of jsBasicGadget
}
\description{
Rstudio addin of jsBasicGadget
}
\details{
Rstudio addin of jsBasicGadget
}
\examples{
if(interactive()){
 jsBasicGadjet(mtcars)
 }
}
\seealso{
\code{\link[rstudioapi]{rstudio-editors}}
}
