% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsBasicGadget.R
\name{jsBasicExtAddin}
\alias{jsBasicExtAddin}
\title{jsBasicExtAddin: RStudio Addin for basic data analysis with external data.}
\usage{
jsBasicExtAddin(nfactor.limit = 20, max.filesize = 2048)
}
\arguments{
\item{nfactor.limit}{nlevels limit for categorical variables, Default: 20}

\item{max.filesize}{Maximum file size to upload (MB), Default: 2048 (2 GB)}
}
\value{
RStudio Addin for basic data analysis with external data.
}
\description{
RStudio Addin for basic data analysis with external csv/xlsx/sas7bdat/sav/dta file.
}
\details{
RStudio Addin for basic data analysis with external csv/xlsx/sas7bdat/sav/dta file.
}
\examples{
if (interactive()) {
  jsBasicExtAddin()
}
}
\seealso{
\code{\link[survival]{lung}}
 \code{\link[data.table]{fwrite}}
 \code{\link[jstable]{opt.tbreg}}
}
