% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jstt.R
\name{jsBasicGadget}
\alias{jsBasicGadget}
\title{jsBasicGadget: Shiny Gadget of Basic Statistics in Medical Research.}
\usage{
jsBasicGadget(data, nfactor.limit = 20)
}
\arguments{
\item{data}{data}

\item{nfactor.limit}{nlevels limit for categorical variables}
}
\value{
Shiny Gadget including Data, Label info, Table 1, Regression(linear, logistic), Basic plot
}
\description{
Shiny Gadget including Data, Label info, Table 1, Regression(linear, logistic), Basic plot
}
\details{
Shiny Gadget including Data, Label info, Table 1, Regression(linear, logistic), Basic plot
}
\examples{
if(interactive()){
 jsBasicGadjet(mtcars)
 }
}
