% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsPropensityGadget.R
\name{jsPropensityAddin}
\alias{jsPropensityAddin}
\title{jsPropensityAddin: Rstudio addin of jsPropensityGadget}
\usage{
jsPropensityAddin()
}
\value{
Rstudio addin of jsPropensityGadget
}
\description{
Rstudio addin of jsPropensityGadget
}
\details{
DETAILS
}
\examples{
 #EXAMPLE1
}
\seealso{
\code{\link[rstudioapi]{rstudio-editors}}
}
