% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsPropensityGadget.R
\name{jsPropensityExtAddin}
\alias{jsPropensityExtAddin}
\title{jsPropensityExtAddin: RStudio Addin for propensity score analysis with external data.}
\usage{
jsPropensityExtAddin(nfactor.limit = 20, max.filesize = 2048)
}
\arguments{
\item{nfactor.limit}{nlevels limit for categorical variables, Default: 20}

\item{max.filesize}{Maximum file size to upload (MB), Default: 2048 (2 GB)}
}
\value{
RStudio Addin for propensity score analysis with external data.
}
\description{
RStudio Addin for propensity score analysis with external csv/xlsx/sas7bdat/sav/dta file.
}
\details{
RStudio Addin for propensity score analysis with external csv/xlsx/sas7bdat/sav/dta file.
}
\examples{
if(interactive()){
 jsPropensityExtAddin()
 }
}
\seealso{
\code{\link[survival]{pbc}}
 \code{\link[data.table]{fwrite}},\code{\link[data.table]{data.table-package}}
 \code{\link[survey]{svydesign}}
 \code{\link[jstable]{opt.tbreg}}
}
