% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsPropensityGadget.R
\name{jsPropensityGadget}
\alias{jsPropensityGadget}
\title{jsPropensityGadget: Shiny Gadget for propensity score analysis.}
\usage{
jsPropensityGadget(data, nfactor.limit = 20)
}
\arguments{
\item{data}{data}

\item{nfactor.limit}{nlevels limit for categorical variables, Default: 20}
}
\value{
Shiny Gadget including original/matching/IPTW data, Label info, Table 1, Cox model, Basic/kaplan-meier plot.
}
\description{
Shiny Gadget including original/matching/IPTW data, Label info, Table 1, Cox model, Basic/kaplan-meier plot.
}
\details{
Shiny Gadget including original/matching/IPTW data, Label info, Table 1, Cox model, Basic/kaplan-meier plot.
}
\examples{
if (interactive()) {
  jsPropensityGadget(mtcars)
}
}
\seealso{
\code{\link[data.table]{data.table}}
 \code{\link[MatchIt]{matchit}},\code{\link[MatchIt]{match.data}}
 \code{\link[jstable]{cox2.display}},\code{\link[jstable]{svycox.display}}
 \code{\link[survival]{survfit}},\code{\link[survival]{coxph}},\code{\link[survival]{Surv}}
 \code{\link[jskm]{jskm}},\code{\link[jskm]{svyjskm}}
 \code{\link[ggplot2]{ggsave}}
 \code{\link[survey]{svykm}}
}
