% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsRepeatedGadget.R
\name{jsRepeatedAddin}
\alias{jsRepeatedAddin}
\title{jsRepeatedAddin: Rstudio addin of jsRepeatedAddin}
\usage{
jsRepeatedAddin()
}
\value{
Rstudio addin of jsRepeatedAddin
}
\description{
Rstudio addin of jsRepeatedAddin
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[rstudioapi]{rstudio-editors}}
}
