% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsRepeatedGadget.R
\name{jsRepeatedGadget}
\alias{jsRepeatedGadget}
\title{jsRepeatedGadjet: Shiny Gadget of Repeated measure analysis.}
\usage{
jsRepeatedGadget(data, nfactor.limit = 20)
}
\arguments{
\item{data}{data}

\item{nfactor.limit}{nlevels limit for categorical variables}
}
\value{
Shiny Gadget including Data, Label info, Table 1, GEE(linear, logistic), Basic plot
}
\description{
Shiny Gadget including Data, Label info, Table 1, GEE(linear, logistic), Basic plot
}
\details{
Shiny Gadget including Data, Label info, Table 1, GEE(linear, logistic), Basic plot
}
\examples{
if (interactive()) {
  jsRepeatedGadjet(mtcars)
}
}
