% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsRepeatedGadget.R
\name{jsRepeatedGadjet}
\alias{jsRepeatedGadjet}
\title{jsRepeatedGadjet: Shiny Gadget of Repeated measure analysis.}
\usage{
jsRepeatedGadjet(data, nfactor.limit = 20)
}
\arguments{
\item{data}{data}

\item{nfactor.limit}{nlevels limit for categorical variables}
}
\value{
Shiny Gadget including Data, Label info, Table 1, GEE(linear, logistic), Basic plot
}
\description{
Shiny Gadget including Data, Label info, Table 1, GEE(linear, logistic), Basic plot
}
\details{
DETAILS
}
\examples{
if(interactive()){
 library(survival)
 jsRepeatedGadjet(lung)
 }
}
