% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsSurveyGadget.R
\name{jsSurveyGadget}
\alias{jsSurveyGadget}
\title{jsSurveyGadget: Shiny Gadget of survey data analysis.}
\usage{
jsSurveyGadget(data, nfactor.limit = 20)
}
\arguments{
\item{data}{data}

\item{nfactor.limit}{nlevels limit for categorical variables}
}
\value{
Shiny Gadget including Data, Label info, Table 1, svyglm, Basic plot
}
\description{
Shiny Gadget including Data, Label info, Table 1, svyglm, Basic plot
}
\details{
DETAILS
}
\examples{
if(interactive()){
 library(survival)
 jsSurveyGadget(lung)
 }
}
