% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kaplan.R
\name{kaplanModule}
\alias{kaplanModule}
\title{kaplanModule: Module for kaplan-meier plot}
\usage{
kaplanModule(input, output, session, data, data_label,
  data_varStruct = NULL, nfactor.limit = 10, design.survey = NULL,
  id.cluster = NULL)
}
\arguments{
\item{input}{input}

\item{output}{output}

\item{session}{session}

\item{data}{reactive data}

\item{data_label}{reactuve data_label}

\item{data_varStruct}{reactive data_varStruct, Default: NULL}

\item{nfactor.limit}{nlevels limit in factor variable, Default: 10}

\item{design.survey}{reactive survey data. default: NULL}

\item{id.cluster}{reactive cluster variable if marginal model, Default: NULL}
}
\value{
kaplanModule
}
\description{
Module for kaplan-meier plot
}
\details{
Module for kaplan-meier plot
}
\examples{
 #EXAMPLE1
}
