% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display2.R
\name{logistic.display2}
\alias{logistic.display2}
\title{logistic.display2: Modified epiDisplay's logistic.display function.}
\usage{
logistic.display2(
  logistic.model,
  alpha = 0.05,
  crude = TRUE,
  crude.p.value = FALSE,
  decimal = 2,
  simplified = FALSE
)
}
\arguments{
\item{logistic.model}{glm object(binomial)}

\item{alpha}{alpha, Default: 0.05}

\item{crude}{crude, Default: TRUE}

\item{crude.p.value}{crude.p.value, Default: FALSE}

\item{decimal}{decimal, Default: 2}

\item{simplified}{simplified, Default: FALSE}
}
\value{
logistic table
}
\description{
Modified epiDisplay's logistic.display function for reactive data.
}
\details{
Modified epiDisplay's logistic.display function for reactive data.
}
\examples{
model1 <- glm(am ~ cyl + disp, data = mtcars, family = binomial)
logistic.display2(model1, crude = TRUE, crude.p.value = TRUE, decimal = 3)
}
