% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regress.R
\name{logisticModule}
\alias{logisticModule}
\title{logisticModule: Module for logistic regression}
\usage{
logisticModule(input, output, session, data, data_label,
  data_varStruct = NULL, nfactor.limit = 10, design.survey = NULL,
  default.unires = T)
}
\arguments{
\item{input}{input}

\item{output}{output}

\item{session}{session}

\item{data}{data}

\item{data_label}{data_label}

\item{data_varStruct}{data_varStruct, Default: NULL}

\item{nfactor.limit}{nlevels limit in factor variable, Default: 10}

\item{design.survey}{survey data. default: NULL}

\item{default.unires}{Set default independent variables using univariate analysis.}
}
\value{
logisticModule
}
\description{
Module for logistic regression
}
\details{
Module for logistic regression
}
\examples{
library(shiny);library(DT);library(data.table);library(jstable)
ui <- fluidPage(
   sidebarLayout(
   sidebarPanel(
     regressModuleUI("logistic")
   ),
   mainPanel(
     DTOutput("logistictable")
   )
 )
)

server <- function(input, output, session) {

  data <- mtcars
  data.label <- jstable::mk.lev(mtcars)

  out_logistic <- callModule(logisticModule, "logistic", data = data, data_label = data.label,
                           data_varStruct = NULL)

  output$logistictable <- renderDT({
    datatable(out_logistic()$table, rownames=T, caption = out_logistic()$caption)
  })
}
}
