% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mk.lev2}
\alias{mk.lev2}
\title{mk.lev2: level generating function}
\usage{
mk.lev2(out.old, out.label)
}
\arguments{
\item{out.old}{raw data}

\item{out.label}{pre-defined label data}
}
\value{
out.label data labels updated
}
\description{
make level for sav files with labels pre defined from SPSS
}
