% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regress.R
\name{mklist}
\alias{mklist}
\title{mklist: function to make variable list lncluding specific variables.}
\usage{
mklist(varlist, vars)
}
\arguments{
\item{varlist}{Original variable list.}

\item{vars}{variable to include.}
}
\value{
variable list lncluding specific variables.
}
\description{
Function to make variable list lncluding specific variables.
}
\details{
Internal function
}
\examples{
 data_varStruct <- list(variable = names(mtcars))
 mklist(data_varStruct, names(mtcars))
}
