% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regress.R
\name{mksetdiff}
\alias{mksetdiff}
\title{mksetdiff: make setdiff varlist}
\usage{
mksetdiff(varlist, vars)
}
\arguments{
\item{varlist}{varlist}

\item{vars}{vars}
}
\value{
Setdiff varlist
}
\description{
make setdiff varlist
}
\details{
Internal function
}
\examples{
 # Example
}
