% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display2.R
\name{regress.display2}
\alias{regress.display2}
\title{regress.display2: modified epiDisplay's regress.display function}
\usage{
regress.display2(regress.model, alpha = 0.05, crude = FALSE,
  crude.p.value = FALSE, decimal = 2, simplified = FALSE)
}
\arguments{
\item{regress.model}{lm object}

\item{alpha}{alpha, Default: 0.05}

\item{crude}{crude, Default: FALSE}

\item{crude.p.value}{crude.p.value, Default: FALSE}

\item{decimal}{decimal, Default: 2}

\item{simplified}{simplified, Default: FALSE}
}
\value{
regress table
}
\description{
regress.display function for reactive data
}
\details{
regress.display function for reactive data
}
\examples{
model1 <- glm(mpg ~ cyl + disp + vs, data = mtcars)
regress.display2(model1, crude = TRUE, crude.p.value = TRUE, decimal = 3)
}
