% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regress.R
\name{regressModuleUI}
\alias{regressModuleUI}
\title{regressModuleUI: ModuleUI for regression}
\usage{
regressModuleUI(id)
}
\arguments{
\item{id}{id}
}
\value{
regressModuleUI
}
\description{
ModuleUI for regression
}
\details{
ModuleUI for regression
}
\examples{
library(shiny);library(DT);library(data.table);library(jstable)
ui <- fluidPage(
   sidebarLayout(
   sidebarPanel(
     regressModuleUI("linear")
   ),
   mainPanel(
     DTOutput("lineartable")
   )
 )
)

server <- function(input, output, session) {

  data <- reactive(mtcars)
  data.label <- reactive(jstable::mk.lev(mtcars))

  out_linear <- callModule(regressModule2, "linear", data = data, data_label = data.label,
                           data_varStruct = NULL)

  output$lineartable <- renderDT({
    datatable(out_linear()$table, rownames=T, caption = out_linear()$caption)
  })
}
}
