% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc.R
\name{rocUI}
\alias{rocUI}
\title{rocUI: shiny module UI for roc analysis}
\usage{
rocUI(id)
}
\arguments{
\item{id}{id}
}
\value{
Shiny module UI for roc analysis
}
\description{
Shiny module UI for roc analysis
}
\details{
Shiny module UI for roc analysis
}
\examples{
library(shiny);library(DT);library(data.table);library(jstable);library(ggplot2);library(pROC)
ui <- fluidPage(
   sidebarLayout(
   sidebarPanel(
     rocUI("roc")
   ),
   mainPanel(
     plotOutput("plot_roc"),
     tableOutput("cut_roc"),
     ggplotdownUI("roc"),
     DTOutput("table_roc")
   )
 )
)

server <- function(input, output, session) {

  data <- reactive(mtcars)
  data.label <- reactive(jstable::mk.lev(data1))

  out_roc <- callModule(rocModule, "roc", data = data, data_label = data.label,
                            data_varStruct = NULL)

  output$plot_roc <- renderPlot({
    print(out_roc()$plot)
  })

  output$cut_roc <- renderTable({
   print(out_roc()$cut)
  })

  output$table_roc <- renderDT({
    datatable(out_roc()$tb, rownames=F, editable = F, extensions= "Buttons",
              caption = "ROC results",
              options = c(jstable::opt.tbreg("roctable"), list(scrollX = TRUE)))
  })
}
}
