% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb1.R
\name{tb1module2}
\alias{tb1module2}
\title{tb1module2: Shiny module Server for Table 1 using reactive data.}
\usage{
tb1module2(input, output, session, data, data_label,
  data_varStruct = NULL, nfactor.limit = 10, design.survey = NULL)
}
\arguments{
\item{input}{input}

\item{output}{output}

\item{session}{session}

\item{data}{reactive data}

\item{data_label}{reactive data label(reactive)}

\item{data_varStruct}{Variable structure list of data, Default: NULL}

\item{nfactor.limit}{maximum factor levels to include, Default: 10}

\item{design.survey}{Survey data. default: NULL}
}
\value{
Shiny module
}
\description{
Shiny module Server for Table 1 using reactive data.
}
\details{
DETAILS
}
\examples{
 #EXAMPLE1
}
