% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb1simple.R
\name{tb1simple}
\alias{tb1simple}
\title{tb1simple: tb1 module for propensity score analysis}
\usage{
tb1simple(input, output, session, data, matdata, data_label,
  data_varStruct = NULL, group_var)
}
\arguments{
\item{input}{input}

\item{output}{output}

\item{session}{session}

\item{data}{data}

\item{matdata}{matdata}

\item{data_label}{data_label}

\item{data_varStruct}{data_varStruct, Default: NULL}

\item{group_var}{group_var}
}
\value{
tb1 module for propensity score analysis
}
\description{
tb1 module for propensity score analysis
}
\details{
tb1 module for propensity score analysis
}
\examples{
 #EXAMPLE1
}
\seealso{
\code{\link[labelled]{var_label}}
 \code{\link[jstable]{CreateTableOneJS}}
 \code{\link[survey]{svydesign}}
 \code{\link[tableone]{svyCreateTableOne}}
}
