% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb1simple.R
\name{tb1simple2}
\alias{tb1simple2}
\title{tb1simple2: tb1 module for propensity score analysis for reactive data}
\usage{
tb1simple2(input, output, session, data, matdata, data_label,
  data_varStruct = NULL, vlist, group_var)
}
\arguments{
\item{input}{input}

\item{output}{output}

\item{session}{session}

\item{data}{reactive data}

\item{matdata}{reactive matdata}

\item{data_label}{reactive data_label}

\item{data_varStruct}{data_varStruct, Default: NULL}

\item{vlist}{variable lists}

\item{group_var}{reactive group_var}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
tb1 module for propensity score analysis
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[jstable]{CreateTableOneJS}}
 \code{\link[survey]{svydesign}}
 \code{\link[tableone]{svyCreateTableOne}}
}
