% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb1simple.R
\name{tb1simpleUI}
\alias{tb1simpleUI}
\title{tb1simpleUI : tb1 module UI for propensity score analysis}
\usage{
tb1simpleUI(id)
}
\arguments{
\item{id}{id}
}
\value{
UI
}
\description{
tb1 module UI for propensity score analysis
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
