% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templateGenerator.R
\name{templateGenerator}
\alias{templateGenerator}
\title{templateGenerator: Shiny Gadget for global/app.R template.}
\usage{
templateGenerator()
}
\value{
Shiny Gadget including Data, Label info, Table 1, Regression(linear, logistic), Basic plots
}
\description{
Opens a Shiny app that allows users to generate a Shiny project template.
}
\details{
Shiny Gadget including Data, Label info, Table 1, Regression(linear, logistic), Basic plots
}
\examples{
if (interactive()) {
  templateGenerator()
}
}
