% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeroc.R
\name{timeROC_table}
\alias{timeROC_table}
\title{timeROC_table: extract AUC information from list of timeROC object.}
\usage{
timeROC_table(ListModel, dec.auc = 3, dec.p = 3)
}
\arguments{
\item{ListModel}{list of timeROC object}

\item{dec.auc}{digits for AUC, Default: 3}

\item{dec.p}{digits for p value, Default: 3}
}
\value{
table of AUC information
}
\description{
extract AUC information from list of timeROC object.
}
\details{
extract AUC information from list of timeROC object.
}
\examples{
#library(survival)
#list.timeROC <- lapply(list("age", c("age", "sex")),
#                      function(x){
#                        timeROChelper("status", "time", x, t = 365, data = lung)
#                       })
#timeROC_table(list.timeROC)
}
\seealso{
\code{\link[stats]{confint}}
 \code{\link[data.table]{data.table-package}}
}
