% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeroc.R
\name{timeROChelper}
\alias{timeROChelper}
\title{timeROChelper: Helper function for timerocModule}
\usage{
timeROChelper(
  var.event,
  var.time,
  vars.ind,
  t,
  data,
  design.survey = NULL,
  id.cluster = NULL
)
}
\arguments{
\item{var.event}{event}

\item{var.time}{time}

\item{vars.ind}{independent variable}

\item{t}{time}

\item{data}{data}

\item{design.survey}{survey data, Default: NULL}

\item{id.cluster}{cluster variable if marginal model, Default: NULL}
}
\value{
timeROC and coxph object
}
\description{
Helper function for timerocModule
}
\details{
Helper function for timerocModule
}
\examples{
#library(survival)
#timeROChelper("status", "time", c("age", "sex"), t = 365, data = lung)
}
\seealso{
\code{\link[survival]{coxph}}
 \code{\link[survey]{svycoxph}}
 \code{\link[stats]{predict}}
 \code{\link[timeROC]{timeROC}}
}
