% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csvFileInput.R
\name{csvFile}
\alias{csvFile}
\title{csvFile: Shiny module Server for file upload.}
\usage{
csvFile(input, output, session, nfactor.limit = 20)
}
\arguments{
\item{input}{input}

\item{output}{output}

\item{session}{session}

\item{nfactor.limit}{nfactor limit to include, Default: 20}
}
\value{
server
}
\description{
Shiny module Server for file(csv or xlsx) upload.
}
\details{
Shiny module Server for file(csv or xlsx) upload.
}
\examples{
library(shiny);library(DT);library(data.table);library(readxl);library(jstable)
ui <- fluidPage(
  sidebarLayout(
    sidebarPanel(
      csvFileInput("datafile")
    ),
    mainPanel(
      tabsetPanel(type = "pills",
                  tabPanel("Data", DTOutput("data")),
                  tabPanel("Label", DTOutput("data_label", width = "100\%"))
                 )
   )
 )
)

server <- function(input, output, session) {
  data <- callModule(csvFile, "datafile")

  output$data <- renderDT({
    data()$data
  })

  output$label <- renderDT({
    data()$label
  })
}
}
