% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csvFileInput.R
\name{csvFile}
\alias{csvFile}
\title{csvFile: Shiny module Server for file upload.}
\usage{
csvFile(input, output, session)
}
\arguments{
\item{input}{input}

\item{output}{output}

\item{session}{session}
}
\value{
server
}
\description{
Shiny module Server for file(csv or xlsx) upload.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 library(shiny)
 ui <- fluidPage(
   sidebarLayout(
     sidebarPanel(
       csvFileInput("datafile", "Upload data (csv/xlsx format)")
     ),
     mainPanel(
       dataTableOutput("table")
     )
   )
 )
 server <- function(input, output, session) {
   datafile <- callModule(csvFile, "datafile")

   output$table <- renderDataTable({
     datafile()
   })
 }

 shinyApp(ui, server)
 }
}
}
