% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilePsInput.R
\name{FilePs}
\alias{FilePs}
\title{FilePs: Shiny module Server for file upload for propensity score matching.}
\usage{
FilePs(input, output, session, nfactor.limit = 20)
}
\arguments{
\item{input}{input}

\item{output}{output}

\item{session}{session}

\item{nfactor.limit}{nfactor limit to include, Default: 20}
}
\value{
server
}
\description{
Shiny module Server for file upload for propensity score matching.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
