% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilePsInput.R
\name{FilePsInput}
\alias{FilePsInput}
\title{FilePsInput: Shiny module UI for file upload for propensity score matching.}
\usage{
FilePsInput(id, label = "Upload data (csv/xlsx/sav/sas7bdat/dta)")
}
\arguments{
\item{id}{id}

\item{label}{label, Default: 'csv/xlsx/sav/sas7bdat file'}
}
\value{
Shiny UI
}
\description{
Shiny module UI for file upload for propensity score matching.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
