% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileSurveyInput.R
\name{FileSurveyInput}
\alias{FileSurveyInput}
\title{FileSurveyInput: File upload UI for survey data analysis.}
\usage{
FileSurveyInput(id, label = "Upload data (csv/xlsx/sav/sas7bdat/dta)")
}
\arguments{
\item{id}{id}

\item{label}{label, Default: 'csv/xlsx/sav/sas7bdat/dta file'}
}
\value{
Shiny UI
}
\description{
File upload UI for survey data analysis.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 library(shiny);library(DT);library(data.table);library(readxl);library(jstable)
 ui <- fluidPage(
   sidebarLayout(
     sidebarPanel(
       FileSurveyInput("datafile")
     ),
     mainPanel(
       tabsetPanel(type = "pills",
                   tabPanel("Data", DTOutput("data")),
                   tabPanel("Label", DTOutput("data_label", width = "100\%"))
                   )
     )
   )
 )

 server <- function(input, output, session) {
   data <- callModule(FileSurvey, "datafile")

   output$data <- renderDT({
     data()$data
   })

   output$label <- renderDT({
     data()$label
   })
 }

 shinyApp(ui, server)
 }
}
}
