% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestcox.R
\name{forestcoxUI}
\alias{forestcoxUI}
\title{forestcoxUI:shiny module UI for forestcox}
\usage{
forestcoxUI(id, label = "forestplot")
}
\arguments{
\item{id}{id}

\item{label}{label, Default: 'forestplot'}
}
\value{
Shinymodule UI
}
\description{
Shiny module UI for forestcox
}
\details{
Shinymodule UI for forestcox
}
\examples{

library(survival);library(dplyr);library(jstable);library(shiny)
lung \%>\%
  mutate(
    status = as.integer(status == 1),
    sex = factor(sex),
    kk = factor(as.integer(pat.karno >= 70)),
    kk1 = factor(as.integer(pat.karno >= 60))

  ) -> lung
out<-lung
ui <- fluidPage(
  sidebarLayout(
    sidebarPanel(
      forestcoxUI('Forest')
    ),
    mainPanel(
      DTOutput('tablesub')
    )
  )
)


server <- function(input, output, session) {
  data<-reactive(out)
 label<-reactive(mk.lev(out))
 outtable<-forestcoxServer('Forest',data=data,data_label=label)
  output$tablesub<-renderDT({
    outtable()
  })
}

shinyApp(ui, server)

}
