% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestcox.R
\name{forestcoxUI}
\alias{forestcoxUI}
\title{forestcoxUI:shiny module UI for forestcox}
\usage{
forestcoxUI(id, label = "forestplot")
}
\arguments{
\item{id}{id}

\item{label}{label, Default: 'forestplot'}
}
\value{
Shinymodule UI
}
\description{
Shiny module UI for forestcox
}
\details{
Shinymodule UI for forestcox
}
\examples{
ui <- fluidPage(
  sidebarLayout(
    sidebarPanel(
      forestcoxUI('Forest')
    ),
    mainPanel(
      DTOutput('tablesub')
    )
  )
)


server <- function(input, output, session) {
  data<-reactive(out)

  outtable<-forestcoxServer('Forest',data=data,data_label=mk.lev(data) )
  output$tablesub<-renderDT({
    outtable()
  })
}
}
