% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestglm.R
\name{forestglmUI}
\alias{forestglmUI}
\title{forestglmUI:Shiny module UI for forestglm}
\usage{
forestglmUI(id, label = "forestplot")
}
\arguments{
\item{id}{id}

\item{label}{label, Default: 'forestplot'}
}
\value{
Shinymodule UI
}
\description{
Shiny module UI for forestcox
}
\details{
Shinymodule UI for forestglm
}
\examples{
\dontrun{
if(interactive()){
library(survival)
library(dplyr)

lung \%>\%
  mutate(
    status = factor(as.integer(status == 1)),
    sex = factor(sex),
    kk = factor(as.integer(pat.karno >= 70)),
    kk1 = factor(as.integer(pat.karno >= 60))
  ) -> lung
library(shiny);library(DT)
ui <- fluidPage(
  sidebarLayout(
    sidebarPanel(
      forestglmUI('Forest')
    ),
    mainPanel(
      DTOutput('tablesub')
    )
  )
)

out<-lung

library(jstable);
library(data.table)
server <- function(input, output, session) {
  data<-reactive(out)
  label<-reactive(mk.lev(out))
  outtable<-forestglmServer('Forest',data=data,data_label=label,family='gaussian')
  output$tablesub<-renderDT({
    outtable()
  })
}

shinyApp(ui, server)

 }
}
}
