% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpairs.R
\name{ggpairsModule}
\alias{ggpairsModule}
\title{ggpairsModule: ggpairs module}
\usage{
ggpairsModule(input, output, session, data, data_label,
  data_varStruct = NULL)
}
\arguments{
\item{input}{input}

\item{output}{output}

\item{session}{session}

\item{data}{data}

\item{data_label}{data_label}

\item{data_varStruct}{data_varStruct, Default: NULL}
}
\value{
ggpairsModule
}
\description{
ggpairs module
}
\details{
ggpairs module
}
\examples{
 #EXAMPLE1
}
