% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpairs.R
\name{ggpairsModule2}
\alias{ggpairsModule2}
\title{ggpairsModule2: ggpairs module for reactive data}
\usage{
ggpairsModule2(input, output, session, data, data_label,
  data_varStruct = NULL)
}
\arguments{
\item{input}{input}

\item{output}{output}

\item{session}{session}

\item{data}{reactive data}

\item{data_label}{reactive data_label}

\item{data_varStruct}{data_varStruct, Default: NULL}
}
\value{
ggpairsModule2
}
\description{
ggpairs module for reactive data
}
\details{
DETAILS
}
\examples{
 #EXAMPLE1
}
