% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpairs.R
\name{ggpairsModuleUI1}
\alias{ggpairsModuleUI1}
\title{ggpairsModuleUI1: Variable selection module UI for ggpairs}
\usage{
ggpairsModuleUI1(id)
}
\arguments{
\item{id}{id}
}
\value{
ggpairsModuleUI1
}
\description{
Variable selection module UI for ggpairs
}
\details{
DETAILS
}
\examples{
 #EXAMPLE1
}
