% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kaplan.R
\name{ggplotdownUI}
\alias{ggplotdownUI}
\title{ggplotdownUI: Option & download module UI for ggplot}
\usage{
ggplotdownUI(id)
}
\arguments{
\item{id}{id}
}
\value{
ggplotdownUI
}
\description{
Option & download module UI for ggplot
}
\details{
DETAILS
}
\examples{
 #EXAMPLE1
}
