% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsPropensityGadget.R
\name{jsPropensityAddin}
\alias{jsPropensityAddin}
\title{jsPropensityAddin: Rstudio addin of jsPropensityGadget}
\usage{
jsPropensityAddin()
}
\value{
Rstudio addin of jsPropensityGadget
}
\description{
Rstudio addin of jsPropensityGadget
}
\details{
Rstudio addin of jsPropensityGadget
}
\examples{
if(interactive()){
 jsPropensityGadget(mtcars)
 }
}
\seealso{
\code{\link[rstudioapi]{rstudio-editors}}
}
