% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsRepeatedGadget.R
\name{jsRepeatedAddin}
\alias{jsRepeatedAddin}
\title{jsRepeatedAddin: Rstudio addin of jsRepeatedGadget}
\usage{
jsRepeatedAddin()
}
\value{
Rstudio addin of jsRepeatedGadget
}
\description{
Rstudio addin of jsRepeatedGadget
}
\details{
Rstudio addin of jsRepeatedGadget
}
\examples{
if (interactive()) {
  jsRepeatedAddin()
}
}
\seealso{
\code{\link[rstudioapi]{rstudio-editors}}
}
