% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsRepeatedGadget.R
\name{jsRepeatedAddin}
\alias{jsRepeatedAddin}
\title{jsRepeatedAddin: Rstudio addin of jsRepeatedGadjet}
\usage{
jsRepeatedAddin()
}
\value{
Rstudio addin of jsRepeatedGadjet
}
\description{
Rstudio addin of jsRepeatedGadjet
}
\details{
Rstudio addin of jsRepeatedGadjet
}
\examples{
if(interactive()){
 jsRepeatedAddin()
 }
}
\seealso{
\code{\link[rstudioapi]{rstudio-editors}}
}
