% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsSurveyGadget.R
\name{jsSurveydAddin}
\alias{jsSurveydAddin}
\title{jsSurveydAddin: Rstudio addin of jsSurveyGadget}
\usage{
jsSurveydAddin()
}
\value{
Rstudio addin of jsSurveyGadget
}
\description{
Rstudio addin of jsSurveyGadget
}
\details{
Rstudio addin of jsSurveyGadget
}
\examples{
if(interactive()){
 jsSurveydAddin()
 }
}
\seealso{
\code{\link[rstudioapi]{rstudio-editors}}
}
