% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kaplan.R
\name{kaplanUI}
\alias{kaplanUI}
\title{kaplanUI: ModuleUI for kaplan-meier plot}
\usage{
kaplanUI(id)
}
\arguments{
\item{id}{id}
}
\value{
kaplanUI
}
\description{
ModuleUI for kaplan-meier plot
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
