% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regress.R
\name{logisticModule}
\alias{logisticModule}
\title{logisticModule: Module for logistic regression}
\usage{
logisticModule(input, output, session, data, data_label,
  data_varStruct = NULL, nfactor.limit = 10, design.survey = NULL)
}
\arguments{
\item{input}{input}

\item{output}{output}

\item{session}{session}

\item{data}{data}

\item{data_label}{data_label}

\item{data_varStruct}{data_varStruct, Default: NULL}

\item{nfactor.limit}{nlevels limit in factor variable, Default: 10}

\item{design.survey}{survey data. default: NULL}
}
\value{
logisticModule
}
\description{
Module for logistic regression
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
