% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regress.R
\name{logisticModule2}
\alias{logisticModule2}
\title{logisticModule2: Module for logistic regression using reactive data}
\usage{
logisticModule2(input, output, session, data, data_label,
  data_varStruct = NULL, nfactor.limit = 10, design.survey = NULL,
  default.unires = T)
}
\arguments{
\item{input}{input}

\item{output}{output}

\item{session}{session}

\item{data}{reactive data}

\item{data_label}{reactive data_label}

\item{data_varStruct}{data_varStruct, Default: NULL}

\item{nfactor.limit}{nlevels limit in factor variable, Default: 10}

\item{design.survey}{survey data. default: NULL}

\item{default.unires}{Set default independent variables using univariate analysis.}
}
\value{
logisticModule2
}
\description{
Module for logistic regression using reactive data
}
\details{
Module for logistic regression using reactive data
}
\examples{
library(shiny);library(DT);library(data.table);library(jstable)
ui <- fluidPage(
   sidebarLayout(
   sidebarPanel(
     regressModuleUI("logistic")
   ),
   mainPanel(
     DTOutput("logistictable")
   )
 )
)

server <- function(input, output, session) {

  data <- reactive(mtcars)
  data.label <- reactive(jstable::mk.lev(mtcars))

  out_logistic <- callModule(logisticModule2, "logistic", data = data, data_label = data.label,
                           data_varStruct = NULL)

  output$logistictable <- renderDT({
    datatable(out_logistic()$table, rownames=T, caption = out_logistic()$caption)
  })
}
}
