% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regress.R
\name{mklist}
\alias{mklist}
\title{mklist: make intersect varlist}
\usage{
mklist(varlist, vars)
}
\arguments{
\item{varlist}{varlist}

\item{vars}{vars}
}
\value{
Intersect varlist
}
\description{
make intersect varlist
}
\details{
Internal function
}
\examples{
\dontrun{
if(interactive()){
 data_varStruct <- list(variable = names(data))
 conti_list <- mklist(data_varStruct, conti_vars)
 }
}
}
