% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regress.R
\name{mksetdiff}
\alias{mksetdiff}
\title{mksetdiff: function to make variable list excluding specific variables.}
\usage{
mksetdiff(varlist, vars)
}
\arguments{
\item{varlist}{Original variable list}

\item{vars}{variable to exclude.}
}
\value{
variable list excluding specific variables.
}
\description{
Function to make variable list excluding specific variables.
}
\details{
Internal function
}
\examples{
 data_varStruct <- list(variable = names(mtcars))
 mksetdiff(data_varStruct, "mpg")
}
