% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regress.R
\name{regressModule}
\alias{regressModule}
\title{regressModule: Module for linear regression}
\usage{
regressModule(input, output, session, data, data_label,
  data_varStruct = NULL)
}
\arguments{
\item{input}{input}

\item{output}{output}

\item{session}{session}

\item{data}{data}

\item{data_label}{data_label}

\item{data_varStruct}{data_varStruct, Default: NULL}
}
\value{
regressModule
}
\description{
Module for linear regression
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
