% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regress.R
\name{regressModule}
\alias{regressModule}
\title{regressModule: Shiny modulde server for linear regression.}
\usage{
regressModule(input, output, session, data, data_label,
  data_varStruct = NULL, nfactor.limit = 10, design.survey = NULL,
  default.unires = T)
}
\arguments{
\item{input}{input}

\item{output}{output}

\item{session}{session}

\item{data}{data}

\item{data_label}{data label}

\item{data_varStruct}{List of variable structure, Default: NULL}

\item{nfactor.limit}{nlevels limit in factor variable, Default: 10}

\item{design.survey}{survey data. default: NULL}

\item{default.unires}{Set default independent variables using univariate analysis, Default: T}
}
\value{
Shiny modulde server for linear regression.
}
\description{
Shiny modulde server for linear regression.
}
\details{
Shiny modulde server for linear regression.
}
\examples{
library(shiny);library(DT);library(data.table);library(jstable)
ui <- fluidPage(
   sidebarLayout(
   sidebarPanel(
     regressModuleUI("linear")
   ),
   mainPanel(
     DTOutput("lineartable")
   )
 )
)

server <- function(input, output, session) {

  data <- mtcars
  data.label <- jstable::mk.lev(mtcars)

  out_linear <- callModule(regressModule, "linear", data = data, data_label = data.label,
                           data_varStruct = NULL)

  output$lineartable <- renderDT({
    datatable(out_linear()$table, rownames=T, caption = out_linear()$caption)
  })
}
}
