% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regress.R
\name{regressModule}
\alias{regressModule}
\title{regressModule: Module for linear regression}
\usage{
regressModule(input, output, session, data, data_label,
  data_varStruct = NULL, nfactor.limit = 10, design.survey = NULL,
  default.unires = T)
}
\arguments{
\item{input}{input}

\item{output}{output}

\item{session}{session}

\item{data}{data}

\item{data_label}{data_label}

\item{data_varStruct}{data_varStruct, Default: NULL}

\item{nfactor.limit}{nlevels limit in factor variable, Default: 10}

\item{design.survey}{survey data. default: NULL}

\item{default.unires}{Set default independent variables using univariate analysis.}
}
\value{
regressModule
}
\description{
Module for linear regression
}
\details{
DETAILS
}
\examples{
 #EXAMPLE1
}
