% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regress.R
\name{regressModule2}
\alias{regressModule2}
\title{regressModule2: Module for linear regression using reactive data}
\usage{
regressModule2(input, output, session, data, data_label,
  data_varStruct = NULL)
}
\arguments{
\item{input}{input}

\item{output}{output}

\item{session}{session}

\item{data}{reactive data}

\item{data_label}{reactive data_label}

\item{data_varStruct}{data_varStruct, Default: NULL}
}
\value{
regressModule2
}
\description{
Module for linear regression using reactive data
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
