% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regress.R
\name{regressModuleUI}
\alias{regressModuleUI}
\title{regressModuleUI: ModuleUI for regression}
\usage{
regressModuleUI(id)
}
\arguments{
\item{id}{id}
}
\value{
regressModuleUI
}
\description{
ModuleUI for regression
}
\details{
DETAILS
}
\examples{
 #EXAMPLE1
}
