% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb1.R
\name{tb1module}
\alias{tb1module}
\title{tb1module: Shiny module Server for Table 1.}
\usage{
tb1module(input, output, session, data, data_label,
  data_varStruct = NULL, nfactor.limit = 10, design.survey = NULL)
}
\arguments{
\item{input}{input}

\item{output}{output}

\item{session}{session}

\item{data}{data}

\item{data_label}{data label}

\item{data_varStruct}{Variable structure list of data, Default: NULL}

\item{nfactor.limit}{maximum factor levels to include, Default: 10}

\item{design.survey}{survey data. default: NULL}
}
\value{
Shiny module
}
\description{
Shiny module Server for Table 1.
}
\details{
DETAILS
}
\examples{
 #EXAMPLE1
}
