% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb1.R
\name{tb1module}
\alias{tb1module}
\title{tb1module: Shiny module Server for Table 1.}
\usage{
tb1module(input, output, session, data, data_label,
  data_varStruct = NULL, nfactor.limit = 10)
}
\arguments{
\item{input}{input}

\item{output}{output}

\item{session}{session}

\item{data}{data}

\item{data_label}{data label}

\item{data_varStruct}{Variable structure list of data, Default: NULL}

\item{nfactor.limit}{maximum factor levels to include, Default: 10}
}
\value{
Shiny module
}
\description{
Shiny module Server for Table 1.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 library(shiny);library(data.table);library(DT)
 library(jstable);library(shinycustomloader);library(tableone);library(labelled)
 data = data.table(mtcars)
 data$vs = as.factor(data$vs)
 data$am = as.factor(data$am)
 data$cyl = as.factor(data$cyl)
 data.label = mk.lev(data)

 ui <- fluidPage(
 sidebarLayout(
   sidebarPanel(
     tb1moduleUI("tb1")
   ),
   mainPanel(
     withLoader(DTOutput("table1"), type="html", loader="loader6"),
   )
 )
 )

 server <- function(input, output, session) {
   out_tb1 <- callModule(tb1module, "tb1",
                         data = data, data_label = data.label, data_varStruct = NULL)
   output$table1 <- renderDT({
     tb = out_tb1()$table
     cap = out_tb1()$caption
     out.tb1 = datatable(tb, rownames = T, extension= "Buttons", caption = cap,
                         options = c(opt.tb1("tb1"),
                                     list(columnDefs = list(list(visible=FALSE, targets= which(colnames(tb) \%in\% c("test","sig"))))
                                     ),
                                     list(scrollX = TRUE)
                         )
     )
     if ("sig" \%in\% colnames(tb)){
       out.tb1 = out.tb1 \%>\% formatStyle("sig", target = 'row' ,backgroundColor = styleEqual("**", 'yellow'))
     }
     return(out.tb1)
   })
 }

 shinyApp(ui, server)
 }
}
}
