% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb1.R
\name{tb1moduleUI}
\alias{tb1moduleUI}
\title{tb1moduleUI: Shiny module UI for Table 1.}
\usage{
tb1moduleUI(id)
}
\arguments{
\item{id}{id}
}
\value{
Shiny UI
}
\description{
Shiny module UI for Table 1.
}
\details{
DETAILS
}
\examples{
 #EXAMPLE1
}
